+++ SupportFE/KiraJewelry/src/main/java/com/jewelry/KiraJewelry/repository/ProductDesignShellRepository.java
@@ -9,6 +9,6 @@ import com.jewelry.KiraJewelry.models.ProductDesignShell;
 public interface ProductDesignShellRepository extends JpaRepository<ProductDesignShell, Integer> {
 
     @Query("SELECT pds FROM ProductDesignShell pds WHERE pds.material_Id = :materialId AND pds.weight = :weight")
-    ProductDesignShell findByMaterialIdAndWeight(@Param("materialId") int materialId, @Param("weight") Float weight);
+    ProductDesignShell findByMaterialIdAndWeight(@Param("materialId") int materialId, @Param("weight") int weight);
 }
